fix(feed): address blocking review feedback - Add per-user enrichment cache (30s TTL) to For You pipeline,   eliminating 6 uncached DB reads per authenticated request - Extract shared useFeed hook from useForYouFeed/useNarrativeFeed,   removing ~400 lines of duplicated state machine code - Fix retry off-by-one in useFeedEventTracker (was allowing 4   attempts instead of 3) - Fix LikeButton handleReactionSelect toggling like state when   changing reaction type on an already-liked post
